change block hold timeout to float (#509)

+++ src/bxcommon/models/blockchain_network_model.py
@@ -17,7 +17,7 @@ class BlockchainNetworkModel:
     block_interval: int
     ignore_block_interval_count: int
     block_recovery_timeout_s: int
-    block_hold_timeout_s: int
+    block_hold_timeout_s: float
     final_tx_confirmations_count: int
     tx_contents_memory_limit_bytes: int
     max_block_size_bytes: int = constants.DEFAULT_MAX_PAYLOAD_LEN_BYTES
